requirements.txt
scapy==2.5.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
joblib==1.3.0
matplotlib==3.7.0
pyyaml==6.0
fastparquet==2023.8.0
